Support older versions of requests package
Support for the `json` parameter was added to requests version 2.4.2,
published in 2014. (https://github.com/requests/requests/blob/master/HISTORY.rst#242-2014-10-05)
We could require a newer `requests` version in setup.py, but we might as
well just be compatible with the older versions that still seem to be out in the
wild.