Add client instances
For beelines, we're likely going to want to initialize libhoney in the beeline in a way that could conflict with how the consumer is already using libhoney. There may also be places in the beeline where we need to init libhoney, but can't know if it's already initialized in other places. It's safer if we can have a client instance that's isolated from what's going on in the global state.