Deprecate send_now method, add send, update docs
`send_now` has been (justifiably) confusing to folks because it implies we stop everything and send the event now, when in fact all we do is enqueue the event to be sent later. We're deprecating this method. It will be removed eventually, but for now we'll update docs to point to alternatives.