[transmission] safely handle sending on pool during shutdown
Since we use a daemon thread to send events, it's possible (especially when libhoney.close() is not called) for application termination to cause early termination of the ThreadPoolExecutor. If we're trying to submit events to it at the time, an unhandled RuntimeError can occur and leak. Since it's happening during application shutdown in an isolated thread, it's not too problematic, but can cause confusion. Worst case scenario, let's just silently discard the error and move on.

Fixes https://github.com/honeycombio/libhoney-py/issues/43